Extract shared Form 34A capture sheet with safe-area fix
Both AddFormModal and CounterEvidenceModal now render one perk-styled
presentational sheet (header, capture button, vote rows, floating footer)
and keep only their own data/submit logic. The modal wraps its content in
its own SafeAreaProvider so the header clears the status bar.
